pyproject.toml
[project]
name = "blog-recommendation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "faiss-cpu>=1.11.0",
    "fastapi>=0.115.14",
    "greenlet>=3.2.3",
    "sacremoses>=0.1.1",
    "sentence-transformers>=5.0.0",
    "sentencepiece>=0.2.0",
    "sqlalchemy>=2.0.41",
    "transformers>=4.53.0",
    "uvicorn>=0.35.0",
]
